refactor: rename postgresql_loader.py to loader.py
