+++ src/langchain_google_cloud_sql_pg/__init__.py
@@ -15,7 +15,7 @@
 from . import indexes
 from .chat_message_history import PostgresChatMessageHistory
 from .engine import Column, PostgresEngine
-from .postgresql_loader import PostgresDocumentSaver, PostgresLoader
+from .loader import PostgresDocumentSaver, PostgresLoader
 from .vectorstore import PostgresVectorStore
 from .version import __version__
